requirements.txt
luma.lcd>=2.5.0
Pillow>=10.0.0
paho-mqtt>=1.6.1
RPi.GPIO>=0.7.1
smbus2>=0.4.3
